Fixed inbound and outbound pairing issues
This change fixes several issues around inbound and outbound pairing.  Inbound pairing was not launching the inbound pairing screen properly.  Inbound pairing handler would crash if multiple inbound pairing requests occurred.  Inbound pairing and Outbound pairing would fail if concurrently in progress (this change separates them).  This change improves scrolling so that onyl the listview scrolls and not the whole screen.  This change also fixes an issues that scrambled PIN handling.
(ConfirmOnly was not asking for user confirmation, and DisplayPin was asking for confirmation).

This change also adds english resources for certain labels that were hard-coded.  It also introduces some new labels.

This change also allows the full width of the available screen to be utilized and changes the text wrapping rules for inbound and outbound pairing so that it can break on lengthy bluetooth ID's if needed.