Allow running commands as admin
Uses WDP REST APIs to provide the ability to run a command using Administrator account. To run as admin, the command should be preceded by "runasadmin". The UI will then prompt for the administrator password.
Prerequisites: WDP must be present and running on the default port and protocol. Additionally, loopback must be allowed for the app. "CheckNetIsolation.exe LoopbackExempt" can be used to enable a loopback exemption.
Also, this commit contains a few other UI fixes and code cleanup.